Comment out
                        while wx.GetApp().Pending():
                            wx.GetApp().Dispatch()
                            wx.GetApp().Yield(True)
blocks.  This just seems to hang the program completely.  It is possible
that this is an attempt to make threading with the GUI work better but
for now it is just hanging.  Also added in some code in exception
handling to try & bring out more info when there are problems.